+++ src/main/java/com/GarageApp/GarageApp/Entity/UserEntity.java
@@ -10,7 +10,7 @@ public class UserEntity {
     @Id
     @Column(name = "id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "user_name", nullable = false)
     private String username;
@@ -34,18 +34,14 @@ public class UserEntity {
     }
 
 
-
-
-
-
     @OneToOne
     @JoinColumn(name = "role_id")
     private RoleEntity roles;
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
